# api/openapi.yml
openapi: 3.0.0

info:
  version: 1.0.0
  title: Training sessions receiver service API
  contact:
    name: Dmitry Novikov
    email: dredfort.42@gmail.com
    url: "https://github.com/dredfort42"
  license:
    name: GNU General Public License v3.0
    url: "https://www.gnu.org/licenses/gpl-3.0.html"
  x-logo:
    url: "https://avatars.githubusercontent.com/u/102029973?v=4"
  description: |
    This is a description of the Training sessions receiver service API and its operations.

tags:
  - name: User
    description: |
      The training session uploaded from the browser.
  - name: Device
    description: |
      The training session uploaded from the device.

servers:
  - url: "http://localhost:8424"
    description: |
      Local server

paths:
  /api/v1/training/session:
    post:
      tags:
        - User
        - Device
      summary: Create a new training session
      description: |
        Create a new user's training session in the training session receiver service.
      security:
        - OAuth2AccessToken: []
      operationId: "TrainingSessionCreate"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainingSession"
      responses:
        "201":
          description: |
            OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Training session created successfully"
                  session_uuid:
                    type: string
                    example: "61c4bfb0-e4b6-485f-9147-732c259bf447"
        "400":
          description: |
            Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                missingToken:
                  summary: Missing token
                  value:
                    error: "invalid_request"
                    error_description: "Missing or invalid token"
                invalidJSON:
                  summary: Invalid JSON
                  value:
                    error: "invalid_request"
                    error_description: "Error binding JSON | "
                invalidEmailInSessionData:
                  summary: Invalid email in session data
                  value:
                    error: "invalid_request"
                    error_description: "Invalid email in session data"
                invalidSessionTime:
                  summary: Invalid session time
                  value:
                    error: "invalid_request"
                    error_description: "Invalid session time"
                invalidSessionData:
                  summary: Error parsing session data
                  value:
                    error: "invalid_request"
                    error_description: "Error parsing session data | ..."
        "500":
          description: |
            Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                missingEmail:
                  summary: Missing email
                  value:
                    error: "invalid_request"
                    error_description: "Missing email"
                createSessionError:
                  summary: Create session error
                  value:
                    error: "server_error"
                    error_description: "Error creating session | ..."

    delete:
      tags:
        - User
      summary: Delete training session
      description: |
        Delete training session from the training session receiver service.
      security:
        - OAuth2AccessToken: []
      operationId: "DeleteSession"
      parameters:
        - name: session_uuid
          in: query
          description: |
            The session UUID to delete
          required: true
          schema:
            type: string
            format: uuid
            example: "61c4bfb0-e4b6-485f-9147-732c259bf447"

      responses:
        "200":
          description: |
            OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Training session deleted successfully"
        "400":
          description: |
            Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                missingToken:
                  summary: Missing token
                  value:
                    error: "invalid_request"
                    error_description: "Missing or invalid token"
                missingSessionUUID:
                  summary: Missing session_uuid
                  value:
                    error: "invalid_request"
                    error_description: "Missing session_uuid"
        "404":
          description: |
            Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                sessionNotFound:
                  summary: Training session not exist
                  value:
                    error: "not_found"
                    error_description: "Training session not exist"
        "500":
          description: |
            Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                missingEmail:
                  summary: Missing email
                  value:
                    error: "invalid_request"
                    error_description: "Missing email"
                deleteSessionError:
                  summary: Delete session error
                  value:
                    error: "server_error"
                    error_description: "Error deleting training session | ..."

components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: access_token

  schemas:
    Error:
      description: |
        Error response
      type: object
      properties:
        error:
          type: string
          example: "invalid_request"
        error_description:
          type: string
          example: Missing required fields

    SessionGeneralData:
      description: |
        Training session general information
      type: object
      properties:
        email:
          type: string
          example: "john.smith@example.com"
        device_name:
          type: string
          example: "My_Device"
        session_start_time:
          type: integer
          example: 1719591677
        session_end_time:
          type: integer
          example: 1719593879
      required:
        - email
        - session_start_time
        - session_end_time

    SessionRouteData:
      description: |
        Training session route data
      type: object
      properties:
        timestamp:
          type: integer
          example: 1719591694
        latitude:
          type: number
          example: 51.504425323023213
        longitude:
          type: number
          example: -0.33814574803019221
        horizontal_accuracy:
          type: number
          example: 11.999369782516258
        altitude:
          type: number
          example: 18.462055553045673
        vertical_accuracy:
          type: number
          example: 40.674209594726562
        speed:
          type: number
          example: 1.2921162843704224
        speed_accuracy:
          type: number
          example: 0.66185456514358521
        course:
          type: number
          example: 269.70616594515167
        course_accuracy:
          type: number
          example: 32.025281697329099

    TrainingSession:
      description: |
        Training session data
      allOf:
        - $ref: "#/components/schemas/SessionGeneralData"
        - type: object
          properties:
            route_data:
              type: array
              items:
                $ref: "#/components/schemas/SessionRouteData"
            step_count:
              type: array
              items:
                example:
                  timestamp: 1719591687
                  quantity: "1732 count"
            running_power:
              type: array
              items:
                example:
                  timestamp: 1719591697
                  quantity: "77 W"
            vertical_oscillation:
              type: array
              items:
                example:
                  timestamp: 1719591717
                  quantity: "9.4 cm"
            energy_burned:
              type: array
              items:
                example:
                  timestamp: 1719591681
                  quantity: "0.198 kcal"
            heart_rate:
              type: array
              items:
                example:
                  timestamp: 1719591684
                  quantity: "1.56667 count/s"
            stride_length:
              type: array
              items:
                example:
                  timestamp: 1719591717
                  quantity: "0.89 m"
            ground_contact_time:
              type: array
              items:
                example:
                  timestamp: 1719591723
                  quantity: "242 ms"
            speed:
              type: array
              items:
                example:
                  timestamp: 1719591697
                  quantity: "0.946416 m/s"
            distance:
              type: array
              items:
                example:
                  timestamp: 1719591687
                  quantity: "6.37101 m"
            vo2_max:
              type: array
              items:
                example:
                  timestamp: 1719593877
                  quantity: "49.17 mL/min·kg"